refactor: Extract menu action item
